+++ CloudShell.ControlPlane.Providers/Runtime/ProviderExecution.cs
@@ -622,6 +622,9 @@ public static class ProviderExecutionInstructionTypes
     public const string JavaScriptAppStart = "javaScriptApp.start";
     public const string JavaScriptAppStop = "javaScriptApp.stop";
     public const string JavaScriptAppRestart = "javaScriptApp.restart";
+    public const string JavaAppStart = "javaApp.start";
+    public const string JavaAppStop = "javaApp.stop";
+    public const string JavaAppRestart = "javaApp.restart";
     public const string ExecutableApplicationStart = "executableApplication.start";
     public const string ProcessStart = "process.start";
     public const string ProcessStop = "process.stop";